Now i'm fully in c#
The past year a lot of things happened.  I couldt'n keep my learning path as python web developer while working and trying to recover my mental health. So I changed my mind to become a C# pro; Yeah, I learned a lot of things about python, but at the end of the day i'm still working with C#, and i probably will be working with C# for the rest of 2023 because i like my team, i like my work , and i like the incoming callenges. So now I want to improve as far as i can to be the best developer i can be, and the language i will use to reach that goal is C#.